jaxrs: update RefundJson with adjusted invoice items
We will add the ability to automatically adjust invoice items
when issuing refunds.
In preparation for that feature, update the refund object to
potentially contain the linked (adjusted) invoice items.

When querying for refunds, we don't return yet the associated invoice
items. We should probably add a query parameter to do it.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>