beatrix: test the overdue notification queue
The test was previously checking the overdue states by explicitly
calling refresh on the bundle (the configuration parameter
autoReevaluationInterval was so high, that the notification queue never
kicked in).

This wasn't fully testing overdue though. For instance, the test
testOverdueStateIfNoPaymentMethod() would pass even if we don't send
the payment errors on the bus when there is no default payment method
(because we were asking overdue to refresh its state manually).
Since there is no payment retry if the account doesn't have a default
payment method, overdue has to rely on the notification queue to refresh
its state (and make the bundle go through the various overdue stages
when we move the clock).

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>