account: fix jdbi exception in AuditedAccountEmailDao
AuditedCollectionDaoBase was failing to notice the difference between existing emails
and new ones because the underlying comparison is flawed (it expects the
same class). This was making server throw a 500 error when trying to
add an existing email address.

We really need to fix AuditedCollectionDaoBase eventually. For now, apply a
band-aid.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>