entitlement: fix transfer issue with migrated bundles
For migrated bundles, the last event is likely to be a MIGRATE_BILLING
event, which is ignored in toEvents() when building the events for the transfered
bundle.
This was causing errors when transferring bundles where the destination
subscription wouldn't have any event.

This patch does best effort to find at least one event by building
them as we go (the last non-null one would be used, in this case the one
associated with the MIGRATE_ENTITLEMENT event).

Another solution would be to create a dummy TRANSFER event by default
although this is a bit tricky to build (need to remember the last phase name,
etc.).

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>