osgi-bundles: introduce slf4j-osgi library
For OSGI bundles relying on third-party libraries using slf4j (e.g.
config-magic), there needs to be a slf4j implementation that bridges
with LogServices. Otherwise, slf4j will default to a no-op
implementation (resulting in the loss of messages).

This new library is designed to bridge that gap. To use it, simply add
it to your OSGI bundle and call:

  StaticLoggerBinder.getSingleton().setLogService(logService);

in the start method of your activator (logService is the
OSGIKillbillLogService provided by KillbillActivatorBase).

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>