Refactor overdue code and introduce a new notificationQ to handle asynchronously all bus events.
Each queue has been abstracted to provide its own logic to delete future notifications.
The new queue (overdue-async-bus-queue) will remove future notifications for a given account
and therefore bring some guarantee -- although this is not a strict guarantee -- that we don't refresh/clear
overdue state for a given account at the same time.

One of the intent of that work was to address issue #94

This should also address #85 until we solve #58

Will close the correcponsing bugs after code review