junction: fix bug when computing blocking durations
This patch removes

  getBlockingAll(UUID blockableId, BlockingStateType blockingStateType, InternalTenantContext context);

from BlockingInternalApi: junction was calling it multiple times for different blockingStateType, which
was breaking the natural ordering of events and hence introducing bugs when computing blocking durations.

Instead, we introduce

  getBlockingAllForAccount(InternalTenantContext context);

Events are ordered correctly by junction and will return bundle, subscription and account level blocking
events (we were ignoring the subscription level ones before for some reason).

This also:

* fixes the ordering returned by the SQL statements (only really needed for getBlockingStateForService
as we re-order the other calls in ProxyBlockingStateDao).
* removes BillingEventSet#isLast which wasn't used.

This closes #123.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>