Fixes #139
Yet another implementation for the convertion back from a LocalDate in account TZ to
fully qualified UTC DateTime, based on the referenceTime.

The implementation simply keeps the offset (number of days) that was substracted/added when we initially
converted the DateTime into Localdate (account TZ), and use that imformation to do the reverse operation.