See #141
This enables all the setters for the ModelDao so that we don't use reflection when mapping back results from ResultSet.
When there is a large number of results returned * large number of columns this starts to become significant.
This will save between 10 and 15 uSec per value, so for 400 rows returned * 15 fields -> 90 mSec.

(If the call makese several of those queries, for e.g 3 queries -> 270 mSec).