osgi-bundles: fix bug in logger bundle registration process
If a LogReaderService was already present when starting the logger
bundle, we were manually registering the KillbillLogWriter using a ServiceTracker,
but calling close on it right away. This was calling LogReaderServiceFactory#ungetService which
in turn was calling LogReaderServiceImpl#removeAllLogListeners (i.e.
removing the Kill Bill log listener).

Revisit that scenario by simply manually constructing a REGISTERED
ServiceEvent and letting our ServiceListener take care of the
registration.

This fixes #100.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>