Invoice fix to address issues related with leading pro-rations.
In the case of billingPeriod different than MONTHLy the current code would do an initial leading pro-ration to the billCycleDate within
the current or next months. Those would cause different issues as documented in #57, #146 and #160.

The new logic will compute the longest leading pro-ration and as a side effect will address those issues.

Fixes #57, Fixes #146, Fixes #160