properties: refactor the way we fetch system properties
This patch removes the calls to System.getProperties() and centralizes
the management of properties around the ConfigSource object (this work
had already been started, but not yet completed). A nice side effect is
that we can now enforce certain properties (e.g. making sure user.timezone=UTC).

This also helps us delegate properties to a Properties file via
-Dorg.killbill.server.properties=/path/to/killbill.properties.

Finally, this also fixes a bug in UriAccessor in the file access path.

We also re-introduced a test dependency on both MySQL and H2 in all modules. While not
strictly needed (could be only activated in the mysql and h2 profiles), it makes
it easier to run tests in the IDE.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>