Export KillbillConfigSource to plugins so they can retrieve their config.
The introduction of KillbillConfigSource along with the mechanism to specify a config files to load Kill Bill properties broke the plugins, which used to access directly the system properties and would be left with the file name for where properties are instead of properties themselves. With this change, KillbillConfigSource is exported and plugin just need to use the ConfigSource interface to access system properties. The drawback is that we add more package dependencies that need to be passed through import-export osgi mechanism.