invoice: fix test failure on jdk 8
On jdk 7, RolledUpUsageForUnitTypesFactory map values are ordered chronologically.
The sort comparison (via RolledUpUsageForUnitTypes#compareTo) is a no-op, as the
end date of the left interval is equal to the start date of the right interval.

On jdk 8, the map values are inversed. We need to try harder in the compareTo method
to actually sort the intervals.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>