entitlement: export events at the subscription level
The main goal of this patch it to expose the subscription events from the Subscription
object. Before, the ordering logic was encapsulated in DefaultSubscriptionBundleTimeline,
as only the SubscriptionBundle object exported the events, across all subscriptions for
the bundle.

The ordering logic has been taken out of DefaultSubscriptionBundleTimeline into two
pieces: SubscriptionEventOrdering (to order base subscription events) and BlockingStateOrdering
(to insert the blocking states ordered). The code has been moved pretty much as-is.

Besides wiring, the timeline test TestDefaultSubscriptionBundleTimeline is untouched
and passes.

In JAX-RS, the main change is in the SubscriptionJson constructor. While before
we had to extract the events for a given subscription from the bundle timeline,
we can now simply retrieve them from the Subscription object itself. This automatically
fixes https://github.com/killbill/killbill/issues/244.

The full bundle timeline is now exported in BundleTimelineJson as well (the old BundleTimelineJson
format wasn't used).

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>