util: revisit optimization of session update query
It is incorrect to assume that only the last access time
attribute is updated. Try to be a bit smarter instead,
and avoid updating the state for each request. The main drawback
of this approach is that the session may timeout too early
(5% of the timeout length in the worst case, i.e. 90s by default).

See d9b1b00e6fcfa3aac6c2290187d19f29a3e6bb9c and
https://github.com/killbill/killbill/issues/326.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>