ddl: use PostgreSQL-friendly syntax for record_id
In MySQL, SERIAL is an alias for BIGINT UNSIGNED NOT NULL AUTO_INCREMENT UNIQUE.

As part of this change, we are now making all record_id BIGINT instead of INT.
We need to specify unique, as it is not automatic in PostgreSQL.

Note that INT will be used in PostgreSQL, not BIGINT (MySQL doesn't support
bigserial). Hopefully, this shouldn't be an issue - we don't expect any table
to have over 4,294,967,295 rows.

The default values and constraints for account_record_id and tenant_record_id
have also been fixed (should never be null).

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>