invoice: Rework next notification date for in arrear usage billing.
The previous code would use the endDate of the latest (in arrear usage) item, and apply some heurstics based on that date
to figureb out the next notification date associated to that specific usage section in a specific subscription.

Instead, we now compute thgat logic from the in arrear usage code where all the information is already available.