Fixes #453
The EntitelementState associated to subscriptions that are started in the future is currently null. The current patch now handles cases
where that state is null to allow for instance operation such as as cancellation (the issue reported). In addition, the computation for the
effectiveDate of such cancellation (bith at the subscription base level and entitlement level) has been modified to make sure
we don't use the current date (clock.getUTCNow()) but instaed we use the subscription startDate (cancelling a subscription prior its startDate
does not make any sense).

Tests will follow in seperate commit to keep it short.