subscription: Change the logic to cancel the future events associated to a subscription when we make a new operation on the subscription (CHANGE, CANCEL,..)
In the current code we identify the events to be cancelled based on the current time (clock.getUTCNow()). However this is incorrect and we should use instead the
effective date of the cancellation (to find all events that exist after that date, anything before should remain).

One test was modified where we change twice the subscription at the exact same time. In that case we expect 2 events for both change (we could have modified the code
to add an equality instead of stricly greater, but this is such an edge case and in the end it does no matter).