entitlement: fix regression when computing effective cancellation date
ClockUtil will in some cases default to clock.getUTCNow() when transforming the local date
into a date time, which can break add-ons cancellations.

This was a known issue but we re-introduced it after the recent reshuffling of the code
to support bulk cancellations.

Relates to https://github.com/killbill/killbill/issues/451.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>