logging: proposal to output accountRecordId and tenantRecordId in logs
This is a proposal to use the Mapped Diagnostic Context (MDC) to print
the accountRecordId and tenantRecordId in logs:

  http://logback.qos.ch/manual/mdc.html

The MDC is cleared after each request and updated each time an internal context
is created.

This also:

* Switches the requestId implementation to using the MDC rather than a Converter
* Adds the X-Forwarded-For header in the logs (useful when Kill Bill is deployed
  behind a load balancer)

If we decide to go ahead with this approach, more work is required to make sure
the MDC is correctly passed through to children threads (e.g. when going to plugins).

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>